refactor: SignUpEvent 관련 코드 리팩토링
- Event 관련 클래스를 별도의 패키지로 분리
- 클래스명 변경 : SignedUpEvent -> SignUpEvent
- 공식계정 유저 식별자 환경변수명 변경
- 불필요한 기본 생성자 제거
- 오토 포맷팅 적용
