+++ src/main/java/today/seasoning/seasoning/user/service/kakao/KakaoLoginService.java
@@ -16,17 +16,17 @@ import today.seasoning.seasoning.user.dto.LoginResult;
 import today.seasoning.seasoning.user.dto.SocialUserProfileDto;
 
 import java.util.Optional;
+import today.seasoning.seasoning.user.event.SignUpEvent;
 
 @Slf4j
 @Service
 @RequiredArgsConstructor
 public class KakaoLoginService {
-    private static final LoginType KAKAO_LOGIN_TYPE = LoginType.KAKAO;
+
     private final ExchangeKakaoAccessToken exchangeKakaoAccessToken;
     private final FetchKakaoUserProfile fetchKakaoUserProfile;
-    private final UserRepository userRepository;
     private final ApplicationEventPublisher eventPublisher;
-
+    private final UserRepository userRepository;
 
     @Transactional
     public LoginResult handleKakaoLogin(String authorizationCode) {
@@ -46,12 +46,11 @@ public class KakaoLoginService {
     }
 
     private LoginInfo handleLogin(SocialUserProfileDto userProfile) {
-        Optional<User> foundUser = userRepository.find(userProfile.getEmail(), KAKAO_LOGIN_TYPE);
+        Optional<User> foundUser = userRepository.find(userProfile.getEmail(), LoginType.KAKAO);
 
         if (foundUser.isEmpty()) {
             User user = userRepository.save(userProfile.toEntity(LoginType.KAKAO));
-            // 공식 계정 친구 추가 이벤트 발생
-            eventPublisher.publishEvent(new SignedUpEvent(user));
+            eventPublisher.publishEvent(new SignUpEvent(user));
             return new LoginInfo(user, true);
         }
         return new LoginInfo(foundUser.get(), false);
